Move enums to a new module to break circular imports
