+++ pyomo/core/base/indexed_component_slice.py
@@ -15,6 +15,7 @@ import itertools
 from pyomo.common import DeveloperError
 from pyomo.common.collections import Sequence
 
+from pyomo.core.base.enums import SortComponents
 from pyomo.core.base.global_set import UnindexedComponent_index
